Rename prefix/argument/parse.rs => macros.rs and fix docs
